fix: 병원 서비스와 컨트롤러 사이 인터페이스 추가
