+++ src/main/java/com/example/petback/hospital/controller/HospitalController.java
@@ -4,7 +4,6 @@ import com.example.petback.common.security.UserDetailsImpl;
 import com.example.petback.hospital.dto.HospitalRequestDto;
 import com.example.petback.hospital.dto.HospitalResponseDto;
 import com.example.petback.hospital.service.HospitalService;
-import com.example.petback.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
